Reorganise GraphQL and Extensions classes
