+++ _config/graphql.yml
@@ -5,12 +5,9 @@ SilverStripe\GraphQL\Manager:
   schemas:
     admin:
       types:
-        LinkType: 'SilverStripe\Link\Type\TypeCreator'
-        LinkDescription: 'SilverStripe\Link\Type\LinkDescriptionTypeCreator'
+        LinkType: SilverStripe\Link\GraphQL\LinkTypeType
+        LinkDescription: SilverStripe\Link\GraphQL\LinkDescriptionType
 
       queries:
-        readLinkTypes: 'SilverStripe\Link\Type\QueryCreator'
-        readLinkDescription: 'SilverStripe\Link\Type\LinkDescriptionQueryCreator'
-
-#      mutations:
-#        createFile: 'SilverStripe\AssetAdmin\GraphQL\CreateFileMutationCreator'
+        readLinkTypes: SilverStripe\Link\GraphQL\LinkTypeQuery
+        readLinkDescription: SilverStripe\Link\GraphQL\LinkDescriptionQuery
