+++ src/Extensions/ModalController.php
@@ -1,16 +1,23 @@
 <?php
 
-namespace SilverStripe\Link;
+namespace SilverStripe\Link\Extensions;
 
 use InvalidArgumentException;
-use SilverStripe\Admin\ModalController;
+use SilverStripe\Admin\ModalController as OwnerController;
 use SilverStripe\Control\HTTPResponse_Exception;
 use SilverStripe\Core\Extension;
 use SilverStripe\Forms\Form;
+use SilverStripe\Link\FormFactory;
 use SilverStripe\Link\Type\Registry;
 use SilverStripe\View\Requirements;
 
-class ModalControllerExtension extends Extension
+/**
+ * Extensions to apply to ModalController so it knows how to handle the DynamicLink action.
+ *
+ * This action receive a link type key and some link data as a JSON string and retrieve a Form Schema for a
+ * specific Link Type.
+ */
+class ModalController extends Extension
 {
     private static $url_handlers = [
         'editorAnchorLink/$ItemID' => 'editorAnchorLink', // Matches LeftAndMain::methodSchema args
@@ -30,7 +37,7 @@ class ModalControllerExtension extends Extension
         // Show link text field if requested
         $linkDataJsonStr = $this->getOwner()->controller->getRequest()->getVar('data');
 
-        /** @var ModalController $owner */
+        /** @var OwnerController $owner */
         $owner = $this->getOwner();
 
         $factory = FormFactory::singleton();
@@ -44,7 +51,12 @@ class ModalControllerExtension extends Extension
         )->loadDataFrom($data);
     }
 
-    private function getContext()
+    /**
+     * Build the context to pass to the Form Link Factory
+     * @return array
+     * @throws HTTPResponse_Exception
+     */
+    private function getContext(): array
     {
         $linkTypeKey = $this->getOwner()->controller->getRequest()->getVar('key');
         if (empty($linkTypeKey)) {
@@ -65,7 +77,11 @@ class ModalControllerExtension extends Extension
         ];
     }
 
-    private function getData()
+    /**
+     * Extract the Link Data out of the Request.
+     * @return array
+     */
+    private function getData(): array
     {
         $data = [];
         $dataString = $this->getOwner()->controller->getRequest()->getVar('data');
