+++ src/GraphQL/LinkDescriptionType.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace SilverStripe\Link\Type;
+namespace SilverStripe\Link\GraphQL;
 
 use GraphQL\Type\Definition\Type as GraphqlType;
 use SilverStripe\GraphQL\TypeCreator as GraphqlTypeCreator;
 use GraphQL\Type\Definition\ResolveInfo;
 
 /**
- *
+ * GraphQL type for serving a Link Description.
  */
-class LinkDescriptionTypeCreator extends GraphqlTypeCreator
+class LinkDescriptionType extends GraphqlTypeCreator
 {
     /**
      * @return array
